Refactor SD miner, integrate LLM miner, and apply community fixes
