+++ mining_core/base/__init__.py
@@ -1,0 +1,4 @@
+from .config import BaseConfig
+from .model_updater import ModelUpdater
+
+__all__ = ['BaseConfig', 'ModelUpdater']
